[refactor] Wrap BlogPage in Suspense for static
